voxel-game-test - added chat

+++ tests/test-voxel-game/src/main/java/pize/tests/voxelgame/client/chunk/mesh/ChunkBuilder.java
@@ -14,7 +14,7 @@ public class ChunkBuilder{
 
     public static final float AO_BRIGHTNESS = 0.7F;
     
-    public static float buildTime;
+    public static double buildTime;
     public static int vertexCount;
     
     private static Chunk chunk;
